Gossyper may star and unstar gossyp
I decided to place the check to see if a piece of gossyp had been starred onto
the gossyp, because making changes across multiple classes when you don't need
to should be avoided.

I considered placing the message for starring on the Gossyp model; but that
would conflate presentation logic with data logic; so I put it in a helper.

* Extract a before filter to load gossyp

I noticed quite a few `Gossyp.find` calls in my routes; so I hit Rubular.com
and wrote a regular expression to find the right Gossyp for the routes with a
gossyp_id

This article helped me bunches:
http://stackoverflow.com/questions/12375820/how-to-match-regex-route-in-sinatra