Gossyper may unstar gossyp
I decided to move the check to see if a piece of gossyp had been starred onto
the gossyp, because making changes across multiple classes when you don't need
to should be avoided.

I considered placing the message for starring on the Gossyp model; but that
would conflate presentation logic with data logic; so I put it in a helper.